+++ cmd/restic/termstatus.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"sync"
 
-	"github.com/restic/restic/internal/ui"
 	"github.com/restic/restic/internal/ui/termstatus"
 )
 
@@ -31,7 +30,7 @@ func setupTermstatus() (*termstatus.Terminal, func()) {
 
 	// use the termstatus for stdout/stderr
 	prevStdout, prevStderr := globalOptions.stdout, globalOptions.stderr
-	globalOptions.stdout, globalOptions.stderr = ui.WrapStdio(term)
+	globalOptions.stdout, globalOptions.stderr = termstatus.WrapStdio(term)
 
 	return term, func() {
 		// shutdown termstatus
