+++ internal/ui/termstatus/stdio_wrapper_test.go
@@ -1,4 +1,4 @@
-package ui
+package termstatus
 
 import (
 	"strings"
